Fixed float math to correct current_LSB_, shunt voltage, current, and bus voltage all are functional now

+++ Resources/Drivers/SG_Drivers/Inc/INA226.hpp
@@ -33,7 +33,7 @@ public:
     HAL_StatusTypeDef GetShuntVoltage();
     HAL_StatusTypeDef GetCurrent();
     HAL_StatusTypeDef GetBusVoltage();
-    HAL_StatusTypeDef SetCalibrationReg(float ohms, uint8_t maxExpectedCurrent);
+    HAL_StatusTypeDef SetCalibrationReg(float ohms, float maxExpectedCurrent);
     HAL_StatusTypeDef Reset();
     HAL_StatusTypeDef SetConfig();
